requirements.txt
django
djangorestframework
djangorestframework-simplejwt
celery
redis
psycopg2-binary
huggingface_hub
python-dotenv
requests
huggingface_hub
transformers
torch
torchvision
torchaudio
pandas
numpy
scikit-learn
matplotlib
